+++ instrumentation/opentelemetry-instrumentation-kafka-python/src/opentelemetry/instrumentation/kafka/package.py
@@ -18,4 +18,4 @@ _instruments_kafka_python = "kafka-python >= 2.0, < 3.0"
 _instruments_kafka_python_ng = "kafka-python-ng >= 2.0, < 3.0"
 
 _instruments = ()
-_instruments_either = (_instruments_kafka_python, _instruments_kafka_python_ng)
+_instruments_any = (_instruments_kafka_python, _instruments_kafka_python_ng)