+++ instrumentation/opentelemetry-instrumentation-psycopg2/src/opentelemetry/instrumentation/psycopg2/package.py
@@ -17,9 +17,9 @@
 _instruments_psycopg2 = "psycopg2 >= 2.7.3.1"
 _instruments_psycopg2_binary = "psycopg2-binary >= 2.7.3.1"
 
-# TODO: maybe add _instruments_either
+# TODO: maybe add _instruments_any
 _instruments = ()
-_instruments_either = (
+_instruments_any = (
     _instruments_psycopg2,
     _instruments_psycopg2_binary,
 )