+++ CHANGELOG.md
@@ -37,11 +37,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3508](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3508))
 - `opentelemetry-instrumentation-pymongo` `aggregate` and `getMore` capture statements support
   ([#3601](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3601))
-- `opentelemetry-instrumentation-fastapi` Utilize instruments_either functionality. TODO MOVE TO NEW VERSION WHEN OUT
+- `opentelemetry-instrumentation-fastapi` Utilize instruments-any functionality. TODO MOVE TO NEW VERSION WHEN OUT
   ([#3612](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3612))
-- `opentelemetry-instrumentation-psycopg2` Utilize instruments_either functionality.
+- `opentelemetry-instrumentation-psycopg2` Utilize instruments-any functionality.
   ([#3612](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3612))
-- `opentelemetry-instrumentation-kafka-python` Utilize instruments_either functionality.
+- `opentelemetry-instrumentation-kafka-python` Utilize instruments-any functionality.
   ([#3612](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3612))
 
 ## Version 1.34.0/0.55b0 (2025-06-04)